refactor: remove /kit subpath
