--- src/kit/index.ts
@@ -1,5 +1,0 @@
-export { defineNitroPreset } from "./preset";
-export { defineNitroModule } from "./module";
-
-export { writeFile, isDirectory } from "./fs";
-export { prettyPath, resolveNitroPath } from "./path";
