+++ src/presets/_nitro/base-worker.ts
@@ -1,4 +1,4 @@
-import { defineNitroPreset } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
 
 const baseWorker = defineNitroPreset(
   {
