+++ src/presets/_nitro/nitro-dev.ts
@@ -1,4 +1,4 @@
-import { defineNitroPreset } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
 
 const nitroDev = defineNitroPreset(
   {
