+++ src/presets/bun/preset.ts
@@ -1,4 +1,4 @@
-import { defineNitroPreset } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
 
 const bun = defineNitroPreset(
   {
