+++ src/presets/cloudflare/preset.ts
@@ -1,5 +1,5 @@
-import { defineNitroPreset } from "nitro/kit";
-import { writeFile } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
+import { writeFile } from "../_utils/fs";
 import type { Nitro } from "nitro/types";
 import { resolve } from "pathe";
 import { unenvCfExternals } from "../_unenv/preset-workerd";
