+++ src/presets/deno/preset-legacy.ts
@@ -1,8 +1,8 @@
 import { builtinModules } from "node:module";
 import MagicString from "magic-string";
 import { findStaticImports } from "mlly";
-import { defineNitroPreset } from "nitro/kit";
-import { writeFile } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
+import { writeFile } from "../_utils/fs";
 import { isAbsolute, resolve } from "pathe";
 
 // nitro/src/rollup/plugin/import-meta.ts
