+++ src/presets/flightcontrol/preset.ts
@@ -1,4 +1,4 @@
-import { defineNitroPreset } from "nitro/kit";
+import { defineNitroPreset } from "../_utils/preset";
 
 const flightControl = defineNitroPreset(
   {
