+++ src/presets/iis/utils.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import { readFile } from "node:fs/promises";
 import { defu } from "defu";
-import { writeFile } from "nitro/kit";
+import { writeFile } from "../_utils/fs";
 import type { Nitro } from "nitro/types";
 import { resolve } from "pathe";
