+++ src/presets/netlify/legacy/preset.ts
@@ -1,5 +1,5 @@
 import { promises as fsp } from "node:fs";
-import { defineNitroPreset } from "nitro/kit";
+import { defineNitroPreset } from "../../_utils/preset";
 import type { Nitro } from "nitro/types";
 import { dirname, join } from "pathe";
 import { deprecateSWR, writeHeaders, writeRedirects } from "./utils";
