+++ src/rollup/plugins/externals.ts
@@ -9,7 +9,7 @@ import {
   parseNodeModulePath,
 } from "mlly";
 import { resolveModuleURL } from "exsolve";
-import { isDirectory } from "nitro/kit";
+import { isDirectory } from "../../core/utils/fs";
 import type { NodeExternalsOptions } from "nitro/types";
 import { dirname, isAbsolute, join, normalize, relative, resolve } from "pathe";
 import type { PackageJson } from "pkg-types";
