+++ test/fixture/tsconfig.json
@@ -15,7 +15,6 @@
       "nitro/runtime/internal": ["./src/runtime/internal"],
       "nitro/runtime/meta": ["./runtime-meta"],
       "nitro/runtime/*": ["../../src/runtime/*"],
-      "nitro/kit": ["../../src/kit"],
       "nitro/meta": ["../../src/meta"],
       "nitro/presets": ["../../src/presets"],
       "nitro/presets/*": ["../../src/presets/*"],
